requirements.txt
opencv-python>=4.8.0.76
cvzone==1.6.1
numpy>=1.24.4
PyYAML==6.0.2
torch==2.0.1
torchvision==0.15.2
mediapipe>=0.9.0
